% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args2env.R
\name{args2env}
\alias{args2env}
\title{Assign call args to an enviroment.}
\usage{
args2env(call, env = caller_env())
}
\arguments{
\item{call}{[call]}

\item{env}{[environment]  Defaults to caller_env()}
}
\value{
\code{args2env}: [NULL]
}
\description{
Assign call args to an enviroment.
}
\examples{
 args2env(DT(1:10))
 x
}
