% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args2list.R
\name{args2list}
\alias{args2list}
\title{Collect all function arguements as a list.}
\usage{
args2list(env = caller_env())
}
\arguments{
\item{env}{[environment]  Defaults to caller_env()}
}
\value{
\code{args2list}: [list] function arguments
}
\description{
This function should be called before any assignments as the function collects the current function environment.
}
\examples{
 fn = function(a, b) { 
 args2list() 
 } 
 fn(a = 1, b = 3) 
}
