% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_has_unnamed_args.R
\name{call_has_unnamed_args}
\alias{call_has_unnamed_args}
\title{Test if a call has unnamed arguements.}
\usage{
call_has_unnamed_args(x)
}
\arguments{
\item{x}{[call]}
}
\value{
\code{call_has_unnamed_args}: [Logical(1)]
}
\description{
Test if a call has unnamed arguements.
}
\examples{
 x = expr(fn(x, y))
 call_has_unnamed_args(x)
 x = expr(fn(x = x, y = y))
 call_has_unnamed_args(x)
}
