% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccat.R
\name{ccat}
\alias{ccat}
\title{print a character vector as r code.}
\usage{
ccat(x)
}
\arguments{
\item{x}{\code{[character]}}
}
\value{
\code{ccat}: \code{[string]}
}
\description{
print a character vector as r code.
}
\examples{
 ccat(x = c('a', 'b', 'c'))
}
