% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_call_names.R
\name{check_call_names}
\alias{check_call_names}
\title{Find call names in an expression.}
\usage{
check_call_names(
  x,
  call_names,
  env = NULL,
  ignore.wrapper = TRUE,
  ignore.assignments = c("=", "<-"),
  ignore.operators = c("+", "-", "/", "*", "==", "!=", ">=", ">", "<", "!"),
  ignore = NULL
)
}
\arguments{
\item{x}{\code{[expr]}}

\item{call_names}{\code{[character]}  NULL is ok.}

\item{env}{\code{[environment]}  NULL is ok.  Defaults to \code{NULL}}

\item{ignore.wrapper}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{ignore.assignments}{\code{[character]}  NULL is ok.  Defaults to \code{c('=', '<-')}}

\item{ignore.operators}{\code{[character]}  NULL is ok.  Defaults to \code{c('+', '-', '/', '*', '==', '!=', '>=', '>', '<', '!')}}

\item{ignore}{\code{[character]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{check_call_names}: \code{[character]} a character vector of call names
}
\description{
Find call names in an expression.
}
\examples{
 x = expr(SUM(x) + MN(y))
 check_call_names(x, 'SUM')
 x = expr({
 x <- SUM(x) + MN(y)
 })
 check_call_names(x, call_names = 'SUM')
 check_call_names(x, call_names = 'SUM', ignore.assignments = '=')
}
