% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_code_usage.R
\name{check_code_usage}
\alias{check_code_usage}
\title{Check code usage based on valid inputs.}
\usage{
check_code_usage(x, valid_vars, valid_functions, ...)
}
\arguments{
\item{x}{\code{[expr]}}

\item{valid_vars}{\code{[character]}}

\item{valid_functions}{\code{[character]}}

\item{...}{additional arguments}
}
\value{
\code{check_code_usage}: \code{[TRUE|error_message]}
}
\description{
Check code usage based on valid inputs.
}
\examples{
 x = expr({
 SUM(y)
 })
 check_code_usage(x, valid_vars = 'y', valid_functions = 'SUM')
 check_code_usage(x, valid_vars = 'c', valid_functions = 'SM')
 check_code_usage(x, valid_vars = 'y', valid_functions = 'SM')
}
