% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_date_format.R
\name{check_date_format}
\alias{check_date_format}
\title{Check if a string is in a valid date format.}
\usage{
check_date_format(x, format = "ymd", null.ok = TRUE)
}
\arguments{
\item{format}{\code{[choice]}  Possible values: \code{c('ymd_h', 'mdy_h', 'dmy_h', 'ymd_hm', 'mdy_hm', 'dmy_hm', 'ymd_hms',  'mdy_hms', 'dmy_hms', 'ym', 'my', 'yq', 'ymd', 'mdy', 'dmy')}.  Defaults to \code{'ymd'}}

\item{null.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}
}
\value{
\code{check_date_format}: \code{[TRUE|error_message]}
}
\description{
Check if a string is in a valid date format.
}
\examples{
 check_date_format('2022-01-01', format = 'ymd')
 check_date_format('20-01-01', format = 'ymd')
 check_date_format('01/01/2022', format = 'ymd')
 check_date_format('01/30/2022', format = 'dmy')
 check_date_format('01/12/2022', format = 'dmy')
 check_date_format('01/12/2022 12:12:10', format = 'dmy')
 check_date_format('01/12/2022 12:12:10', format = 'dmy_hms')
}
