% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_json.R
\name{check_json}
\alias{check_json}
\title{Check that a string is a valid JSON object.}
\usage{
check_json(x, check_output, null.ok = FALSE)
}
\arguments{
\item{check_output}{\code{[call]}}

\item{null.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{check_json}: \code{[TRUE|error]}
}
\description{
Check that a string is a valid JSON object.
}
\examples{
 check_json(x = '[1,3]', check_numeric(len = 2))
 check_json(x = '[1,3]', check_numeric(len = 1))
 check_json(x = '{\'one\':[1,3]}', check_named_list(structure = list(one = numeric())))
 check_json(x = '{\'one\':[1,3]}', check_named_list(structure = list(two = numeric())))
}
