% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_named_list.R
\name{check_named_list}
\alias{check_named_list}
\title{Check list names and structure.}
\usage{
check_named_list(x, structure, null.ok = FALSE, extra.items = FALSE)
}
\arguments{
\item{x}{\code{[object]} an object to check}

\item{structure}{\code{[list]}  Must be a named list.}

\item{null.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{extra.items}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{check_named_list}: \code{[TRUE|character(message)]}
}
\description{
Check list names and structure.
}
\examples{
 check_named_list(list(x = c('c', NA_character_), y = 1:5),
 structure = list(x = char(), y = numeric()))
 check_named_list(list(x = c('c', NA_character_), y = 1:5),
 structure = list(x = char(any.missing = FALSE), y = numeric()))
 check_named_list(list(x = c('c', NA_character_)), structure = list(x = char(),
 y = numeric()))
 check_named_list(list(x = c('c', NA_character_)), structure = list(x = char(),
 y = numeric(NULL)))
 check_named_list(list(x = LETTERS), structure = list(x = subset(choices = LETTERS)))
 check_named_list(list(x = LETTERS), structure = list(x = subset(choices = letters)))
 check_named_list(list(x = 1, y = 1, z = 4), structure = list(x = num(),
 y = num(), z = num()))
 check_named_list(list(x = 1, y = 1, z = 4, a = 1), structure = list(x = num(),
 y = num(), z = num()))
 check_named_list(list(x = 1, y = 1, z = 4, a = 1), structure = list(x = num(),
 y = num(), z = num()), extra.items = TRUE)
 check_named_list(list(x = 1, y = 1, a = 1), structure = list(x = num(),
 y = num(), z = num()), extra.items = TRUE)
 check_named_list(list(x = 1, y = 1, a = 1), structure = list(x = num(),
 y = num(), z = num(NULL)), extra.items = TRUE)
 check_named_list(NULL, structure = list(x = num(), y = num(),
 z = num(NULL)))
 check_named_list(NULL, structure = list(x = num(), y = num(),
 z = num(NULL)), null.ok = TRUE)
}
