% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_not_empty.R
\name{check_not_empty}
\alias{check_not_empty}
\title{Check if an an empty string, NULL, NA,or length 0.}
\usage{
check_not_empty(x, max_len = NULL, message = "Must have a value")
}
\arguments{
\item{max_len}{\code{[integerish]}  Must have an exact length of \code{1}.  NULL is ok.  Defaults to \code{NULL}}

\item{message}{\code{[string]}  Defaults to \code{'Must have a value'}}
}
\value{
\code{check_not_empty}: \code{[TRUE,error_message]}
}
\description{
Check if an an empty string, NULL, NA,or length 0.
}
\examples{
 check_not_empty(x = '', message = 'Must have a value')
 check_not_empty(x = NULL, message = 'Must have a value')
 check_not_empty(x = character(), message = 'Must have a value')
 check_not_empty(x = NA, message = 'Must have a value')
 check_not_empty(x = c('valid', 'sd'), max_len = 1)
 check_not_empty(x = c('valid'), max_len = 1)
}
