% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_not_in.R
\name{check_not_in}
\alias{check_not_in}
\title{Check to see if a name is not already present.}
\usage{
check_not_in(
  x,
  subset,
  empty.ok = FALSE,
  fail_message = "Must be a unique name.  Name already exists",
  empty_message = "Name is required"
)
}
\arguments{
\item{subset}{\code{[character]}}

\item{empty.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}

\item{fail_message}{\code{[string]}  Defaults to \code{'Must be a unique name.  Name already exists'}}

\item{empty_message}{\code{[character]}  Defaults to \code{'Name is required'}.}
}
\value{
\code{check_not_in}: \code{[TRUE,error_message]}
}
\description{
Check to see if a name is not already present.
}
\examples{
 check_not_in(x = 'Jim', subset = c('Jim', 'Bob'), empty.ok = FALSE)
 check_not_in(x = '', subset = c('Jim', 'Bob'), empty.ok = FALSE)
 check_not_in(x = '', subset = c('Jim', 'Bob'), empty.ok = TRUE)
 check_not_in(x = 'Tim', subset = c('Jim', 'Bob'), empty.ok = FALSE)
}
