% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_undefined_variables.R
\name{check_undefined_variables}
\alias{check_undefined_variables}
\title{Check code for defined variable.}
\usage{
check_undefined_variables(x, init_vars, env = NULL)
}
\arguments{
\item{x}{\code{[expr]}}

\item{init_vars}{\code{[character]}  NULL is ok.}

\item{env}{\code{[environment]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{check_undefined_variables}: \code{[TRUE|class('error_messagae')]} returns TRUE if valid (no undefined variables), otherwise an chararter string of class 'error message'
}
\description{
Check code for defined variable.
}
\examples{
 x = expr({
 x = y * 2
 if (x == 1) {
 out = 1
 }
 else if (x >= 1) {
 1
 }
 else {
 NULL
 }
 out + 1
 })
 check_undefined_variables(x, init_vars = 'y')
 check_undefined_variables(x, init_vars = NULL)
 x = expr({
 x = y * x
 x
 })
 check_undefined_variables(x, init_vars = NULL)
 check_undefined_variables(x, init_vars = 'x')
 check_undefined_variables(x, init_vars = 'y')
 check_undefined_variables(x, init_vars = c('y', 'x'))
 x = expr({
 x = y * x
 c
 })
 check_undefined_variables(x, init_vars = c('y', 'x'))
 check_undefined_variables(x, init_vars = c('y', 'x', 'c'))
 env(x = 1, y = 3)
 check_undefined_variables(x, init_vars = NULL, env = env(x = 1,
 y = 3))
 check_undefined_variables(x, init_vars = NULL, env = env(x = 1,
 y = 3))
}
