% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_code.R
\name{check_valid_code}
\alias{check_valid_code}
\title{Check if a character string is parsable R code.}
\usage{
check_valid_code(x, null.ok = FALSE)
}
\arguments{
\item{x}{\code{[string]}  NULL is ok.}

\item{null.ok}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{FALSE}}
}
\value{
\code{check_valid_code}: \code{[TRUE|character(message)]}
}
\description{
Check if a character string is parsable R code.
}
\examples{
 x = 'SUM(X)'
 check_valid_code('SUM(X)', null.ok = FALSE)
 check_valid_code('SUMX)', null.ok = FALSE)
 check_valid_code('SUM(X){', null.ok = FALSE)
 check_valid_code('SUM(X){}', null.ok = FALSE)
 check_valid_code('SUM(X){\'}', null.ok = FALSE)
 check_valid_code('SUM(X{\']', null.ok = FALSE)
}
