% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_assignment_to_list.R
\name{expr_assignment_to_list}
\alias{expr_assignment_to_list}
\title{Transform expressions with assignments to a named list of expressions}
\usage{
expr_assignment_to_list(x)
}
\arguments{
\item{x}{an expr or list of exprs with assignment/s}
}
\value{
a named list of expressions
}
\description{
Transform expressions with assignments to a named list of expressions
}
\examples{
f <- (c=f+1) ~ (b=f+1) + f(d)
x<-expr_get_formula_args(f)
expr_assignment_to_list(x)
}
