% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_build_formula.R
\name{expr_build_formula}
\alias{expr_build_formula}
\title{Build a formula with character vectors or symbols}
\usage{
expr_build_formula(lhs, rhs)
}
\arguments{
\item{lhs}{left hand side of the formula.  Can be a list of symbols or a character vector.}

\item{rhs}{right hand side of the formula.  Can be a list of symbols or a character vector.}
}
\value{
an expr formula
}
\description{
Build a formula with character vectors or symbols
}
\examples{
lhs <- c('a','b','c')
rhs <- c('x','y','z')
expr_build_formula(lhs,rhs)
}
