% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_call_args.R
\name{expr_call_args}
\alias{expr_call_args}
\title{Extract arguments from an unevaluated call}
\usage{
expr_call_args(x)
}
\arguments{
\item{x}{an call expr}
}
\value{
a list of exprs
}
\description{
Extract arguments from an unevaluated call
}
\examples{
 x=expr(list(mpg=4+1, cyl=4, am="g", vs="f"))
 expr_call_args(x)
}
