% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_call_modify.R
\name{expr_call_modify}
\alias{expr_call_modify}
\title{Modify arguments in a call.}
\usage{
expr_call_modify(call_expr, ...)
}
\arguments{
\item{call_expr}{[call]}

\item{...}{Named expressions (constants, names or calls) used to modify the call. Use zap() to remove arguments. Empty arguments are preserved.}
}
\value{
[Call]
}
\description{
Use zap() to remove an argument.
}
\examples{
 fn = function(a, b, c, d = 0, e = 0) {
 return(a + b + c + d + e)
 }
 call = expr(fn(1, 3, 4))
 expr_call_modify(call, a = 10)
 expr_call_modify(call, d = 5)
 expr_call_modify(call, f = 3)
 fn = function(a, b, c, d = 0, e = 0, ...) {
 return(a + b + c + d + e)
 }
 call_expr = expr(fn(1, 3, 4))
 expr_call_modify(call_expr, f = 3, a = zap())
}
