% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_concat.R
\name{expr_concat}
\alias{expr_concat}
\title{Combine arguments of an expur.}
\usage{
expr_concat(x, y)
}
\arguments{
\item{x}{\code{[call]}}

\item{y}{\code{[call]}  Must be a call from function \code{call_name(x)}.}
}
\value{
\code{expr_concat}: \code{[expr]}
}
\description{
Combine arguments of an expur.
}
\examples{
 x = expr({
 x = 3
 })
 y = expr({
 y = 4
 })
 expr_concat(y, x)
 x = expr(list(x = 3))
 y = expr(list(y = 3))
 expr_concat(y, x)
}
