% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_concat_list_call.R
\name{expr_concat_list_call}
\alias{expr_concat_list_call}
\title{Concatenate two list calls}
\usage{
expr_concat_list_call(x, y)
}
\arguments{
\item{x}{an unevaluated list call}

\item{y}{an unevaluated list call}
}
\value{
concatentated unevaluated list call
}
\description{
Concatenate two list calls
}
\examples{
x=expr(list(mpg,cyl))
y=expr(list(am,vs))
expr_concat_list_call(x,y)
}
