% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_count_calls.R
\name{expr_count_calls}
\alias{expr_count_calls}
\title{Count call names in an expr.}
\usage{
expr_count_calls(x, call_name)
}
\arguments{
\item{x}{\code{[expr]}}

\item{call_name}{\code{[string]}}
}
\value{
\code{expr_count_calls}: [integer(1)]
}
\description{
Count call names in an expr.
}
\examples{
 x = expr(fn(fn(x, y = fn(fn(x)))))
 expr_count_calls(x, call_name = 'fn')
 expr_count_calls(x, call_name = 'x')
}
