% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_deparse_lines.R
\name{expr_deparse_lines}
\alias{expr_deparse_lines}
\title{deparse multiple lines of R code.}
\usage{
expr_deparse_lines(x)
}
\arguments{
\item{x}{\code{[call]}  Must be a bracketed expression.}
}
\value{
\code{expr_deparse_lines}: [character]
}
\description{
deparse multiple lines of R code.
}
\examples{
 expr_deparse_lines({
 sq <- function(x) {
 x^2
 }
 sq(2)
 sq(4)
 })
}
