% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_detect_name.R
\name{expr_detect_name}
\alias{expr_detect_name}
\title{Detect a name in an expr.}
\usage{
expr_detect_name(x, name, search_in_strings = FALSE)
}
\arguments{
\item{x}{\code{[expr]} an r expression.}

\item{name}{\code{[character]}}

\item{search_in_strings}{\code{[logical]}  Defaults to \code{FALSE}}
}
\value{
\code{expr_detect_name}: Logical(1)
}
\description{
Detect a name in an expr.
}
\examples{
 x <- expr(glue_data(dt, 'Hello {name}, you are {get_age(name)} years old.'))
 expr_detect_name(x, name = 'glue_data', search_in_strings = FALSE)
 expr_detect_name(x, name = 'glue_data', search_in_strings = TRUE)
 expr_detect_name(x, name = 'get_age', search_in_strings = FALSE)
 expr_detect_name(x, name = 'get_age', search_in_strings = TRUE)
}
