% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_extract_agg_func.R
\name{expr_extract_agg_func}
\alias{expr_extract_agg_func}
\title{Extract the first aggregration function name from an R expr}
\usage{
expr_extract_agg_func(
  x,
  fill = NA,
  agg_funcs = c("N", "N_missing", "Nth", "N_distinct", "N_duplicated", "first", "last",
    "IQR", "mad", "mean", "median", "max", "min", "sum", "sd", "var", "cor", "kurtosis",
    "skewness", "moment")
)
}
\arguments{
\item{x}{an R expr}

\item{fill}{a name to return if agg call is not present or is ignored}
}
\value{
a call name
}
\description{
Extract the first aggregration function name from an R expr
}
\examples{
x=exprs(sum_mpg=sum(mpg,na.rm=FALSE),mean_mpg=mean(mpg,na.rm=),max_mpg=max(mpg),gg=shift(mpg))
expr_extract_agg_func(x)
expr_set_agg_func(x) \%>\% expr_extract_agg_func()
}
