% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_extract_lines.R
\name{expr_extract_lines}
\alias{expr_extract_lines}
\title{Extract R code lines as a list of exprs.}
\usage{
expr_extract_lines(x)
}
\arguments{
\item{x}{[call]  Must be a bracketed expression.}
}
\value{
\code{expr_make_lines}: [exprs]
}
\description{
Extract R code lines as a list of exprs.
}
\examples{
 expr_extract_lines({
 x = 1:10
 lok <- data.table(x = x)
 lok[, mean(x)]
 })
}
