% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_extract_named_bracket.R
\name{expr_extract_named_bracket}
\alias{expr_extract_named_bracket}
\title{Extract named bracket calls from an expr}
\usage{
expr_extract_named_bracket(
  x,
  name,
  skip_first = TRUE,
  grep = FALSE,
  fully_chained = FALSE
)
}
\arguments{
\item{x}{an expr}

\item{name}{the name to search for}

\item{skip_first}{Logical. should the first named bracket be skipped. Defaults to TRUE}

\item{grep}{Logical. should the name be a pattern to grep}
}
\value{
all named bracket calls matching the name
}
\description{
Extract named bracket calls from an expr
}
\examples{
x=expr(data[,.(.SD[,sum(mpg)]),by=.(am)])
name = ".SD"
expr_extract_named_bracket(x,name)
x=expr(data[,.(.SD_543[,sum(mpg)]),by=.(am)])
name = ".SD"
expr_extract_named_bracket(x,name,grep=FALSE)
expr_extract_named_bracket(x,name,grep=TRUE)
}
