% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_assign.R
\name{expr_find_assign}
\alias{expr_find_assign}
\title{find variables created by assignment.}
\usage{
expr_find_assign(x)
}
\arguments{
\item{x}{\code{[expr]}}
}
\value{
\code{expr_find_assign}: \code{[character]} a character vector of names
}
\description{
find variables created by assignment.
}
\examples{
 expr_find_assign(expr({
 fn <- function(x) {
 out = switch_expr(x, constant = , symbol = character(),
 pairlist = flat_map_chr(x, find_assign_rec),
 call = find_assign_call(x))
 out2 = unique(out)
 out2
 }
 }))
}
