% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_call_names.R
\name{expr_find_call_names}
\alias{expr_find_call_names}
\title{Find call names in an expression.}
\usage{
expr_find_call_names(x, ignore.wrapper = TRUE, ignore = NULL)
}
\arguments{
\item{x}{\code{[expr]}}

\item{ignore.wrapper}{\code{[logical]}  Must have an exact length of \code{1}.  Defaults to \code{TRUE}}

\item{ignore}{\code{[character]}  NULL is ok.  Defaults to \code{NULL}}
}
\value{
\code{expr_find_call_names}: \code{[character]} a character vector of call names
}
\description{
Find call names in an expression.
}
\examples{
 x = expr({
 fd <- function(x) {
 SUM(x) + SUM(y)
 }
 })
 expr_find_call_names(x)
 expr_find_call_names(x, ignore.wrapper = FALSE)
 expr_find_call_names(x, ignore = c('<-', '+'))
}
