% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_constants.R
\name{expr_find_constants}
\alias{expr_find_constants}
\title{Extract constants from an R expression.}
\usage{
expr_find_constants(x, type = NULL)
}
\arguments{
\item{x}{[expr]}

\item{type}{[character]  NULL is ok.  Defaults to NULL}
}
\value{
\code{expr_find_constants}: [vector] a vector of type matches
}
\description{
Extract constants from an R expression.
}
\examples{
 x = expr(list(1L, T, 'a', 1.04))
 expr_find_constants(x)
 expr_find_strings(x)
 expr_find_numeric(x)
 expr_find_integers(x)
 expr_find_logical(x)
}
