% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_replace.R
\name{expr_find_replace}
\alias{expr_find_replace}
\title{Find and replace symbol names in an expr}
\usage{
expr_find_replace(find, replacement, in_expr, match.first = FALSE)
}
\arguments{
\item{find}{a constant, symbol, expr, or exprs}

\item{replacement}{the replacement constant, symbol, expr, or exprs}

\item{in_expr}{the target expression}
}
\value{
an expression
}
\description{
Find and replace symbol names in an expr
}
\examples{
 find = expr(J)
 in_expr = expr(data[, .(J)])
 replacement = exprs(sum_mpg = sum(mpg), max_mpg = max(mpg))
 expr_find_replace(find, replacement, in_expr)

 in_expr = expr(list(list(J),list(J)))
 replacement = exprs(sum_mpg = sum(mpg), max_mpg = max(mpg),min_mpg=min(mpg))
 expr_find_replace(find, replacement, in_expr, match.first = FALSE)
 expr_find_replace(find, replacement, in_expr, match.first = TRUE)
}
