% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_replace_all.R
\name{expr_find_replace_all}
\alias{expr_find_replace_all}
\title{Find and replace multiple matches in an R expr.}
\usage{
expr_find_replace_all(find, replacement, in_expr, match.first = FALSE)
}
\arguments{
\item{find}{[exprs] What to find in the expr.}

\item{replacement}{[exprs]  Must have an exact length of find.}

\item{in_expr}{[expr]}
}
\value{
\code{expr_find_replace_all}: [expr(1)]
}
\description{
Find and replace multiple matches in an R expr.
}
\examples{
 in_expr = expr(data[, .(J), by = by])
 find = exprs(data, J, by)
 replacement = exprs(mtcars, sum(mpg), .(cyl, vs))
 expr_find_replace_all(find, replacement, in_expr)
 in_expr = expr(data[, .(sum(mpg),sum(mpg),sum(mpg))])
 find = exprs(sum(mpg),sum(mpg),sum(mpg))
 replacement =  exprs(sum(V1),sum(V2),sum(V3))
 expr_find_replace_all(find, replacement, in_expr,match.first = FALSE)
 expr_find_replace_all(find, replacement, in_expr,match.first = TRUE)
}
