% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_formula_args_to_list.R
\name{expr_formula_args_to_list}
\alias{expr_formula_args_to_list}
\title{Extract the arguments from a formula}
\usage{
expr_formula_args_to_list(x)
}
\arguments{
\item{x}{a formula expr}
}
\value{
a named list of expressions
}
\description{
Extract the arguments from a formula
}
\examples{
x <- (c=f+1) ~ (b=r+1) + (a=x-1)
expr_formula_args_to_list(x)
expr_get_formula_args(x)
expr_simplify_formula(x)
x= employeeid + (dow = day_of_week(startdtm)) ~ category
expr_get_formula_args(x)
expr_formula_args_to_list(x)
}
