% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_glue.R
\name{expr_glue}
\alias{expr_glue}
\title{Glue a string and parse output into a list of r expressions.}
\usage{
expr_glue(x, env = caller_env())
}
\arguments{
\item{x}{[character|language]  Must have an exact length of 1.}

\item{env}{[environment]  Defaults to caller_env()}
}
\value{
\code{expr_glue}: [list(exressions)]
}
\description{
Glue a string and parse output into a list of r expressions.
}
\examples{
 x = paste0('v', 1:5)
 y = 1:5
 expr_glue('{x}<-{y}')
 #can be in expr format as well.
 expr_glue({x}<-{y})
}
