% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_extract_call.R, R/expr_has_call_name.R
\name{expr_has_call_name}
\alias{expr_has_call_name}
\alias{expr_extract_call}
\title{Extract calls from an expr}
\usage{
expr_extract_call(x, call_name, grep = FALSE, skip_first = TRUE)

expr_has_call_name(x, call_name, skip_first = TRUE)
}
\arguments{
\item{x}{an expr}

\item{call_name}{the call or function name}

\item{grep}{should calls be matched based on a regex pattern? Defaults to false}

\item{skip.first}{\code{[logical(1)]} should the first call be skipped?}
}
\value{
\code{[list(exprs)]} all calls matching the function name. Note: This function does not search recursively.

Logical TRUE/FALSE
}
\description{
Extract calls from an expr

Test if an expr has a call name
}
\examples{
x=expr(data[,.(sum(mpg))])
call_name = "sum"
expr_extract_call(x,call_name)
x=expr(data[,.(sum(mpg))])
call_name = "sum"
expr_has_call_name(x,call_name)

call_name = "mean"
expr_has_call_name(x,call_name)
}
