% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_has_name.R
\name{expr_has_name}
\alias{expr_has_name}
\title{Find variable names in an expr}
\usage{
expr_has_name(x, name, ignore.first = T)
}
\arguments{
\item{x}{an expr}

\item{name}{a symbol name in character format}

\item{ignore.first}{logical. Should the first symbol name be ingored?}
}
\value{
Logical TRUE/FALSA
}
\description{
Find variable names in an expr
}
\examples{
x=expr(data[,.(sum(mpg),mean(vs)),by=.(am)])
expr_has_name(x,name='mpg')
}
