% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_has_named_bracket.R
\name{expr_has_named_bracket}
\alias{expr_has_named_bracket}
\title{Test if an expr contains a named bracket call}
\usage{
expr_has_named_bracket(x, name, skip_first = TRUE, grep = FALSE)
}
\arguments{
\item{x}{an expr}

\item{name}{the name to search for}

\item{skip_first}{Logical. should the first named bracket be skipped. Defaults to TRUE}

\item{grep}{Logical. should the name be a pattern to grep}
}
\value{
Logical TRUE/FALSE
}
\description{
Test if an expr contains a named bracket call
}
\examples{
x=expr(data[,.(.SD[,sum(mpg)]),by=.(am)])
name = ".SD"
expr_has_named_bracket(x,name)
name = "data"
expr_has_named_bracket(x,name)
name = "data"
expr_has_named_bracket(x,name,skip_first=FALSE)
}
