% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_insert.R
\name{expr_insert}
\alias{expr_insert}
\title{Insert an expr into an expr.}
\usage{
expr_insert(x, y, index)
}
\arguments{
\item{x}{\code{[expr]}}

\item{y}{\code{[expr]}}

\item{index}{\code{[integerish]}  NA values are not permitted.}
}
\value{
\code{expr_insert}: [expr]
}
\description{
Insert an expr into an expr.
}
\examples{
 x = expr(list(c('a', 'b', 'c'), 1:5, TRUE))
 y <- expr(unlist)
 expr_insert(x, y, c(1))
 y <- expr(LETTERS[1:3])
 expr_insert(x, y, c(2))
 expr_insert(x, y, c(2, 4))
 expr_insert(x,c(2, 4))<-y
 y
}
