% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_modify_fn_args.R
\name{expr_modify_fn_args}
\alias{expr_modify_fn_args}
\title{Modify functional arguments in an R expression}
\usage{
expr_modify_fn_args(
  x,
  new_args,
  call_names = NULL,
  skip = ".",
  modify_if_present = TRUE
)
}
\arguments{
\item{x}{an R expr}

\item{new_args}{a name list of args to add to specified calls}

\item{call_names}{a character vector indicating the name of the calls (or functions) which will be modified in the expression. If NULL, any call or function with a matching argument name present in the 'new_args' list will be modified, unless a skip is specified.}

\item{skip}{any call names to skip during the search}

\item{modify_if_present}{Logical. If FALSE, the argument will only be modified if it is missing.  See examples}
}
\value{
an expr with updated functional arguments
}
\description{
Modify functional arguments in an R expression
}
\examples{
x=expr(data[,.(sum_mpg=sum(mpg,na.rm=FALSE),mean_mpg=mean(mpg,na.rm=),max_mpg=max(mpg)),by=.(vs)])
expr_modify_fn_args(x,new_args=list(na.rm=TRUE),call_name='sum')
expr_modify_fn_args(x=x,new_args=list(na.rm=TRUE))
expr_modify_fn_args(x=x,new_args=list(na.rm=TRUE),modify_if_present = FALSE)
}
