% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_call_name_args.R
\name{expr_name_call_args}
\alias{expr_name_call_args}
\title{Adds names to unnamed argments in a call and puts arguments in correct order.}
\usage{
expr_name_call_args(x)
}
\arguments{
\item{x}{[call]}
}
\value{
[call] with all args named.
}
\description{
Adds names to unnamed argments in a call and puts arguments in correct order.
}
\examples{
fn=function(a,b,c,d=0,e=0){
   return(a+b+c+d+e)
}
call=expr(fn(1,3,4))
expr_name_call_args(call)
call=expr(fn(d=1,e=3,4,3,1))
expr_name_call_args(call)
call=expr(fn(d=1,e=3,a=4,c=3,b=1))
expr_name_call_args(call)
}
