% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_pluck.R
\name{expr_pluck}
\alias{expr_pluck}
\title{Pluck a piece out of an expr.}
\usage{
expr_pluck(x, index)
}
\arguments{
\item{x}{\code{[expr]}}

\item{index}{\code{[integer]} the index to pluck. \code{NA} values not permitted.}
}
\value{
\code{expr_pluck}: [expr]
}
\description{
Pluck a piece out of an expr.
}
\examples{
 x = expr(list(c('a', 'b', 'c'), 1:5, TRUE))
 expr_pluck(x, c(1))
 expr_pluck(x, c(2))
 expr_pluck(x, c(3))
 expr_pluck(x, c(2, 4))
 expr_pluck(x, c(3, 1))
 expr_pluck(x, c(3, 5))
}
