% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_reduce.R
\name{expr_reduce}
\alias{expr_reduce}
\title{Build code to reduce a list to a single value by iteratively applying a binary function.}
\usage{
expr_reduce(x, func = `\%>\%`)
}
\arguments{
\item{x}{[list] a list of unevaluated expression}

\item{func}{[function]  Defaults to \code{\%>\%}}
}
\value{
An R expression
}
\description{
Build code to reduce a list to a single value by iteratively applying a binary function.
}
\examples{

x<-exprs(t,u,v,x,y,z)
expr_reduce(x,`+`)
}
