% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_remove_parenthesis.R
\name{expr_remove_parenthesis}
\alias{expr_remove_parenthesis}
\title{Remove outer parenthesis from an expr}
\usage{
expr_remove_parenthesis(x)
}
\arguments{
\item{x}{a expr}
}
\value{
a expr with parenthesis removed
}
\description{
Remove outer parenthesis from an expr
}
\examples{
x <- expr((((f+1))))
expr_remove_parenthesis(x)
}
