% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_replace.R
\name{expr_replace}
\alias{expr_replace}
\title{Find and replace code in an r expression.}
\usage{
expr_replace(x, find, replacement)
}
\arguments{
\item{x}{\code{[expr]}}

\item{find}{\code{[expr]}}

\item{replacement}{\code{[any]}}
}
\value{
\code{expr_replace}: [expr]
}
\description{
Find and replace code in an r expression.
}
\examples{
 x = expr(data[, .(x, 'x', x, na.rm = TRUE)])
 expr_replace(x, find = 'x', replacement = expr(y))
 expr_replace(x, find = expr(x), replacement = expr(y))
 expr_replace(x, find = expr(x), replacement = exprs(z, t))
 expr_replace(x, find = expr(x), replacement = exprs(1, 1))
 expr_replace(x, find = TRUE, replacement = FALSE)
}
