% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_replace_call.R
\name{expr_replace_call}
\alias{expr_replace_call}
\title{Replace call names in an expression.}
\usage{
expr_replace_call(x, call_name, replacement, first.only = FALSE)
}
\arguments{
\item{x}{\code{[expr]}}

\item{call_name}{\code{[string]}}

\item{replacement}{\code{[string]}}

\item{first.only}{\code{[logical]}  Defaults to \code{FALSE}}
}
\value{
\code{expr_replace_call}: [expr]
}
\description{
Replace call names in an expression.
}
\examples{
 x = expr(fn(fn(x)))
 expr_replace_call(x, call_name = 'fn', replacement = 'new_fn',
 first.only = TRUE)
 expr_replace_call(x, call_name = 'fn', replacement = 'new_fn',
 first.only = FALSE)
 x = expr({
 x = fn(x)
 fn(fn(x))
 fn
 })
 expr_replace_call(x, call_name = 'fn', replacement = 'new_fn',
 first.only = FALSE)
}
