% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_to_json.R
\name{expr_to_json}
\alias{expr_to_json}
\title{Covert an R expr into a json string.}
\usage{
expr_to_json(x)
}
\arguments{
\item{x}{\code{[call]}}
}
\value{
\code{expr_to_json}: \code{[string]}
}
\description{
Covert an R expr into a json string.
}
\examples{
 expr_to_json(toJSON(out, null = 'null', auto_unbox = T))
}
