% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_to_str.R
\name{expr_to_str}
\alias{expr_to_str}
\title{Convert an expr into a character string.}
\usage{
expr_to_str(x)
}
\arguments{
\item{x}{[expr] an r expression}
}
\value{
\code{expr_to_str}: [Character(1)]
}
\description{
Convert an expr into a character string.
}
\examples{
 x <- expr(x + y + z)
 expr_to_string(x)
}
