% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exprs_eval.R
\name{exprs_eval}
\alias{exprs_eval}
\title{Evaluate a list of expression.}
\usage{
exprs_eval(x, env = caller_env())
}
\arguments{
\item{x}{[list] a list of expressions}

\item{env}{[environment]  Defaults to caller_env()}
}
\value{
\code{exprs_eval}: [NULL]
}
\description{
Evaluate a list of expression.
}
\examples{
 x <- exprs(x <- 3, y <- 4, z <- x + y)
 exprs_eval(x)
}
