% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_make_names.R
\name{exprs_make_names}
\alias{exprs_make_names}
\title{Generate names for a list of expressions}
\usage{
exprs_make_names(x, duplicates.ok = FALSE)
}
\arguments{
\item{x}{list of exprs}

\item{duplicates.ok}{[Logical(1)] Should duplicate names be allowed? Default is FALSE}
}
\value{
a named list of expressions
}
\description{
Generate names for a list of expressions
}
\examples{
x=exprs(sum(hours),sum(hours),max(hours),some_func(hours))
exprs_make_names(x)
x=exprs(sum(x),sum(y),max(y),mean(x),mean(y),mean(y))
exprs_make_names(x,duplicates.ok = FALSE)
}
