% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_lhs_nms.R
\name{f_lhs_nms}
\alias{f_lhs_nms}
\title{Extract the names of a formula's lhs.}
\usage{
f_lhs_nms(f)
}
\arguments{
\item{f}{[formula]}
}
\value{
\code{f_lhs_nms}: [vector(character)]
}
\description{
Extract the names of a formula's lhs.
}
\examples{
 formula = a + b + c ~ x + y + z
 f_lhs_nms(formula)
}
