% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_nms.R
\name{f_nms}
\alias{f_nms}
\title{Extract the names in a formula.}
\usage{
f_nms(f)
}
\arguments{
\item{f}{[formula]}
}
\value{
\code{f_nms}: [vector(character)]
}
\description{
Extract the names in a formula.
}
\examples{
 formula = a + b + c ~ x + y + z
 f_nms(formula)
}
