% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_rhs_nms.R
\name{f_rhs_nms}
\alias{f_rhs_nms}
\title{Extract the names of a formula's lhs.}
\usage{
f_rhs_nms(f)
}
\arguments{
\item{f}{[formula]}
}
\value{
\code{f_rhs_nms}: [vector(character)]
}
\description{
Extract the names of a formula's lhs.
}
\examples{
 formula = a + b + c ~ x + y + z
 f_rhs_nms(formula)
}
