% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_find_names.R
\name{find_names_in_expr}
\alias{find_names_in_expr}
\alias{expr_find_names}
\title{Find variable names in an expr}
\usage{
expr_find_names(x)
}
\arguments{
\item{x}{an expr}
}
\value{
all variable names as class symbol
}
\description{
Find variable names in an expr
}
\examples{
x=expr(data[,.(sum(mpg),mean(vs)),by=.(am)])
expr_find_names(x)
}
