% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_call_name.R
\name{fn_call_name}
\alias{fn_call_name}
\title{Get the name of the current function call.}
\usage{
fn_call_name(n = 1)
}
\arguments{
\item{n}{\code{[integerish]}  Must have an exact length of \code{1}.  Defaults to \code{1}}
}
\value{
\code{fn_call_name}: \code{[character(1)]}
}
\description{
Get the name of the current function call.
}
