% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_assignment.R
\name{is_assignment}
\alias{is_assignment}
\title{Test is an object is an assignment call.}
\usage{
is_assignment(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_assignment}: [Logical(1)]
}
\description{
Test is an object is an assignment call.
}
\examples{
 x = list(1, 2, 4)
 is_assignment(x)
 x <- expr(c(x, 1))
 is_assignment(x)
 x <- expr(x <- 1)
 is_assignment(x)
 x <- expr(`=`(x,1))
 is_assignment(x)
}
