% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_call_logical_test.R
\name{is_call_logical_test}
\alias{is_call_logical_test}
\title{Test if an expr is a logical test.}
\usage{
is_call_logical_test(x)
}
\arguments{
\item{x}{[expr]}
}
\value{
\code{is_call_logical_test}: [Logical(1)]
}
\description{
Test if an expr is a logical test.
}
\examples{
 x = expr(list(x))
 is_call_logical_test(x)
 x = expr(!x)
 is_call_logical_test(x)
 x = expr(x == y)
 is_call_logical_test(x)
 x=  missing_arg()
 is_call_logical_test(x)
}
