% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_expr_equal.R
\name{is_expr_equal}
\alias{is_expr_equal}
\title{Check if two expression are equal.}
\usage{
is_expr_equal(x, y)
}
\arguments{
\item{x}{[expr] an R expression}

\item{y}{[expr] an R expression}
}
\value{
\code{is_expr_equal}: [logical(1)]
}
\description{
Check if two expression are equal.
}
\examples{
 x = expr({
 x = 1
 })
 y = expr({
 x <- 1
 })
 x = expr(function(x) {
 out = lapply(x, function(x) x * 1)
 out
 })
 y = expr(function(x) {
 out = lapply(x, function(y) y * 1)
 out
 })
 is_expr_equal(x, y)
}
