% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_if_statement.R
\name{is_if_statement}
\alias{is_if_statement}
\title{Test if an object is a 'if' statement call.}
\usage{
is_if_statement(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_if_statement}: [Logical(1)]
}
\description{
Test if an object is a 'if' statement call.
}
\examples{
 x <- expr(if (i > 10) {
 print(i)
 })
 is_if_statement(x)
}
