% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_list_call.R
\name{is_list_call}
\alias{is_list_call}
\title{Test if an expr is a list call}
\usage{
is_list_call(x)
}
\arguments{
\item{x}{an object}
}
\value{
Logical TRUE/FALSE
}
\description{
Test if an expr is a list call
}
\examples{
x=expr(list(mpg=4+1, cyl=4, am="g", vs="f"))
is_list_call(x)
x=expr(list())
is_list_call(x)
x=expr(sum(gg))
is_list_call(x)
}
