% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_loop.R
\name{is_loop}
\alias{is_loop}
\title{Test if an object is a for loop call.}
\usage{
is_loop(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_loop}: [Logical(1)]
}
\description{
Test if an object is a for loop call.
}
\examples{
 x <- expr(for (i in 1:10) {
 print(x)
 })
 is_loop(x)
 is_special_call(x)
}
