% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_raw_function.R
\name{is_raw_function}
\alias{is_raw_function}
\title{Test if an object is a raw function call.}
\usage{
is_raw_function(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_raw_function}: [Logical(1)]
}
\description{
Test if an object is a raw function call.
}
\examples{
 x <- expr(function(x) {
 x + 1
 })
 is_raw_function(x)
}
