% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_return.R
\name{is_return}
\alias{is_return}
\title{Test if an object is a return call.}
\usage{
is_return(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_return}: [Logical(1)]
}
\description{
Test if an object is a return call.
}
\examples{
 x = 1
 is_return(x)
 x = expr(c(1, 2, 3))
 is_return(x)
 x = expr(return(y))
 is_return(x)
}
