% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_special_call.R
\name{is_special_call}
\alias{is_special_call}
\title{Test if an object is a raw 'function','if' statement, or 'for' loop call.}
\usage{
is_special_call(x)
}
\arguments{
\item{x}{an object}
}
\value{
\code{is_special_call}: [Logical(1)]
}
\description{
Test if an object is a raw 'function','if' statement, or 'for' loop call.
}
\examples{
 x <- expr(function(x) {
 x + 1
 })
 is_special_call(x)
 x <- expr(for (i in 1:10) {
 print(x)
 })
 is_special_call(x)
 x <- expr(if (i > 10) {
 print(i)
 })
 is_special_call(x)
 x <- expr(glue('This is {x}'))
 is_special_call(x)
}
