% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_expr.R
\name{make_expr}
\alias{make_expr}
\title{Make an r expression from a obj.}
\usage{
make_expr(x)
}
\value{
\code{make_expr}: expr
}
\description{
Make an r expression from a obj.
}
\examples{
 x = data.frame(x = 1:3)
 make_expr(x)
 x = list(1, 3, 4)
 make_expr(x)
}
