% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_names_unique.R
\name{make_names_unique}
\alias{make_names_unique}
\title{Make a character vector (of names) unique.}
\usage{
make_names_unique(x)
}
\arguments{
\item{x}{[character] a character vector, usually names.}
}
\value{
\code{make_names_unique}: [character(length(x))]
}
\description{
Make a character vector (of names) unique.
}
\examples{
x=c('x','x', 'x(2)',NA, 'V1', 'V2', '', 'x(2)',"y","y","V", "x(2)")
make_names_unique(x)
any(duplicated(make_names_unique(x)))
}
