% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_exists.R
\name{package_exists}
\alias{package_exists}
\title{Check if a package exist.}
\usage{
package_exists(pkg)
}
\arguments{
\item{pkg}{[string]}
}
\value{
\code{package_exists}: [Logical(1)]
}
\description{
Check if a package exist.
}
\examples{
 package_exists('data.table')
}
