% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_as_object_name.R
\name{str_as_object_name}
\alias{str_as_object_name}
\title{Turn an string into a valid R object name.}
\usage{
str_as_object_name(x)
}
\arguments{
\item{x}{\code{[character]}  NULL is ok.}
}
\value{
\code{str_as_object_name}: \code{[character]}
}
\description{
Turn an string into a valid R object name.
}
\examples{
 str_as_object_name(x = c('sdf', 'sf', 'safd`', 'sdf a ',
 '123sd', 'sadf{}sa', 'sdf_sdf', 'faf(1)', 'fds2', 'dsf.fs',
 'sdf$fds'))
}
