% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v_collect.R
\name{v_collect}
\alias{v_collect}
\title{Collects the names of the inputs into a function.}
\usage{
v_collect(env = caller_env())
}
\arguments{
\item{env}{[environment]  Defaults to caller_env()}
}
\value{
\code{v_collect}: NULL
}
\description{
Useful for passing variable names into error messages. Variable names are assigned to the function environment as .{arg_name}
}
\examples{
 example_fn <- function(x, y, z) {
 v_collect()
 print(.x)
 print(.y)
 print(.z)
 }
 example_fn(a, b, c)
 example_fn(V1, V2, V3)
 example_fn(1, 2, 3)
}
