% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_named_list.R
\name{assert_named_list}
\alias{assert_named_list}
\title{Assert list names and structure.}
\usage{
assert_named_list(
  x,
  structure,
  null.ok = FALSE,
  extra.items = FALSE,
  env = caller_fn()
)
}
\arguments{
\item{null.ok}{\code{[logical]}  Defaults to \code{FALSE}}

\item{extra.items}{\code{[logical]}  Defaults to \code{FALSE}}

\item{env}{\code{[function]}  Defaults to \code{caller_fn()}}
}
\value{
\code{assert_named_list}: invisible(x)
}
\description{
Assert list names and structure.
}
