% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr_replace_call_names.R
\name{expr_replace_call_names}
\alias{expr_replace_call_names}
\title{Find and replace call names in an expression.}
\usage{
expr_replace_call_names(expr, call_name, replacement)
}
\arguments{
\item{expr}{[expr] an expession}

\item{call_name}{[character] a call name or names}

\item{replacement}{[character]  Must have an exact length of or equal to one of the following: [1,length(call_names)].}
}
\value{
\code{expr_replace_call_names}: [R Expression]
}
\description{
Find and replace call names in an expression.
}
